Set up server directory and user before copying sources to save layer space
Previously, the sources were copied into the container before setting up the
permissions in the server folder which caused all of the sources to be duplicated
in a image layer. To get rid of this extra layer we can set up the permissions and
user first and then copy in the sources.