Properly shut down server after a 'docker stop'
'docker stop' appears to only send a signal to the initial process
started inside of the container which means that the signal will not
make it to child processes without some extra work. Instead, we can
exec both into the entrypoint.sh and into java which means that the
minecraft server will be the same process as the initial process.

I thought the server would stop with a SIGINT since it stops when
you send a CTRL+C to it on an attached terminal, but it appears that
it expects SIGTERM instead for a clean shutdown.